Okay, so I found a SUPER dumb bug.  Logic bug. I had assumed that a certain function, regardless of parameters, after being called, can be 'cached' for its results.  Thats a very dumb assumption.
So now I'm "shoping" the idea of doing an md5sum.
So this idea sorta works, but right now its really slow, because of doing md5sum everytime.  This is probably not the best way to do it.  So lets now try to md5sum this in a much much smarter way.